Create a security package and move TlsUtils.java to it.
